Added vertex AO
+ Remade some textures

+++ VGame/VGame/Header/Util/Util.h
@@ -50,10 +50,10 @@ public:
 
 private:
 	static bool _isInRange(float yaw, int numberToCheckPos, int numberToCheckNeg) {
-		float minPos = numberToCheckPos - CARDIAL_POINT_RANGE,
-			  maxPos = numberToCheckPos + CARDIAL_POINT_RANGE;
-		float minNeg = numberToCheckNeg - CARDIAL_POINT_RANGE,
-			  maxNeg = numberToCheckNeg + CARDIAL_POINT_RANGE;
+		int minPos = numberToCheckPos - CARDIAL_POINT_RANGE,
+			maxPos = numberToCheckPos + CARDIAL_POINT_RANGE;
+		int minNeg = numberToCheckNeg - CARDIAL_POINT_RANGE,
+			maxNeg = numberToCheckNeg + CARDIAL_POINT_RANGE;
 
 		return ((yaw >= minPos) && (yaw <= maxPos) || (yaw >= minNeg) && (yaw <= maxNeg));
 	}
